feat: give every row an owner, so friends can share one instance
Beans, shots, setups, recommendations and settings each gain an `owner`
column and every query is scoped to it, so two people on the same instance
never see or inform each other's coffee.

Identity comes from `core.auth`, which has two modes. `single` -- the
default, and what the personal instance runs -- maps every request to one
implicit owner, leaving behaviour exactly as it was. `tailscale` reads the
`Tailscale-User-Login` header that Tailscale Serve sets from the
authenticated tailnet identity. That header is only trustworthy while the
process is unreachable except through Serve, so `warn_if_misconfigured()`
says so loudly at startup and the compose port stays on loopback.

Migration 0005 backfills every pre-existing row to the placeholder owner
`owner`, which is also the `single` default -- so an existing single-user
database keeps working untouched.

The two CLI entry points have no request to take an identity from, so they
default to the single-user owner: `wcda-backtest` gains `--owner` for
pointing at a multi-user database, and `ai.rag` (the legacy shim behind
`core.main`) resolves one the same way.

+++ src/core/web_schemas.py
@@ -15,7 +15,10 @@ class FeedbackRequest(BaseModel):
     # substituted with a default, because calibration reads these as truth.
     yield_g: float | None = None
     water_g: float | None = None
-    time_s: int | None = None
+    # A float, even though the column is whole seconds: the wizard's timer
+    # reports tenths, and an int here rejected every timed shot outright.
+    # Rounding to the second happens on the way into the row.
+    time_s: float | None = None
     taste_axis: (
         Literal["very_sour", "sour", "balanced", "bitter", "very_bitter"] | None
     ) = None
@@ -112,7 +115,9 @@ class LogDetailsInput(BaseModel):
     grind_setting: str | None = None
     dose_g: float | None = None
     yield_g: float | None = None
-    time_s: int | None = None
+    # Float for the same reason as FeedbackRequest.time_s: a typed "27.5"
+    # must not be a validation error.
+    time_s: float | None = None
     rating: int | None = None
     tasting_notes: str | None = None
